+++ src/main/java/com/smart/booking/domain/payment/service/PaymentHistoryServiceImpl.java
@@ -4,6 +4,7 @@ import com.smart.booking.domain.payment.dto.SavePaymentHistoryDto;
 import com.smart.booking.domain.payment.entity.PaymentHistory;
 import com.smart.booking.domain.payment.repositroy.PaymentHistoryRepository;
 import com.smart.booking.domain.payment.repositroy.PaymentRepository;
+import java.math.BigDecimal;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,7 +20,7 @@ public class PaymentHistoryServiceImpl implements PaymentHistoryService {
         paymentHistoryRepository.save(
             PaymentHistory.builder()
                 .payment(historyDto.payment())
-                .totalAmount(historyDto.totalAmount())
+                .totalAmount(BigDecimal.valueOf(historyDto.totalAmount()))
                 .paymentStatus(historyDto.paymentStatus())
                 .build()
         );
